requirements.txt
pymupdf 
pdfplumber 
reportlab 
pytesseract 
pillow